feat: add `sueParser` helper
- added `sueParser` to reduce code intensity in `listItemParser`

SUE-2
